requirements.txt
aioredis==1.3.1
appnope==0.1.0
argon2==0.1.10
argon2-cffi==20.1.0
asgiref==3.3.1
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
autobahn==19.3.3
Automat==0.7.0
awscli==1.19.27
awsebcli==3.19.4
backcall==0.2.0
bcrypt==3.2.0
blessed==1.15.0
boto3==1.17.27
botocore==1.20.27
CacheControl==0.12.6
cached-property==1.4.3
cachetools==3.1.0
cement==2.8.2
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.6
coreschema==0.0.4
cryptography==2.8
daphne==2.2.2
decorator==4.1.2
Django==3.1.5
django-appconf==1.0.3
django-braces==1.12.0
django-cleanup==4.0.0
django-crispy-forms==1.7.0
django-debug-toolbar==3.2
django-filter==1.1.0
django-otp==1.0.6
django-resized==0.3.9
django-rest-knox==4.1.0
django-sslserver==0.20
django-storages==1.10
django-utils-six==2.0
django-webpack-loader==1.0.0
django-websocket-redis==0.5.2
djangorestframework==3.11.0
docker==4.4.4
docker-compose==1.25.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
ffmpeg-python==0.2.0
firebase-admin==4.3.0
fsspec==0.4.1
future==0.16.0
gevent==1.3.6
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==0.4.14
grpcio==1.36.1
hiredis==0.2.0
httpie==0.9.9
httplib2==0.15.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.10.2
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
Markdown==2.6.11
MarkupSafe==1.1.1
msgpack==0.5.6
names==0.3.0
packaging==20.9
paramiko==2.7.1
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.4
Pillow==7.2.0
prompt-toolkit==3.0.14
proto-plus==1.17.0
protobuf==3.12.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.5.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
Pygments==2.2.0
PyHamcrest==1.9.0
pylibmc==1.6.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.7.3
python-decouple==3.3
pytz==2017.2
PyYAML==5.3.1
qrcode==7.1
redis==2.10.6
requests==2.25.1
rsa==3.4.2
s3transfer==0.3.4
semantic-version==2.8.5
service-identity==17.0.0
shutilwhich==1.1.0
simplegeneric==0.8.1
six==1.13.0
sqlparse==0.3.0
stick-protocol-server==1.2.10
stripe==7.0.0
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
traitlets==4.3.2
Twisted==18.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.4
uWSGI==2.0.17.1
virtualenv==15.1.0
wcwidth==0.1.7
websocket-client==0.49.0
zope.interface==5.1.0
